refactor: Move client related data structure to data.client
