+++ src/winter/controller/FrontController.java
@@ -16,10 +16,10 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletRequestWrapper;
 import jakarta.servlet.http.HttpServletResponse;
 import winter.controller.FrontController;
-import winter.data.JsonString;
 import winter.data.Mapping;
 import winter.data.MappingMethod;
-import winter.data.ModelView;
+import winter.data.client.JsonString;
+import winter.data.client.ModelView;
 import winter.data.enumdata.RequestVerb;
 import winter.exception.AnnotationNotFoundException;
 import winter.exception.DuplicateMappingException;
