+++ src/winter/data/client/ModelView.java
@@ -1,4 +1,4 @@
-package winter.data;
+package winter.data.client;
 
 import java.util.HashMap;
 import java.util.Map;
